@@ -13,24 +13,21 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@babel/core": "^7.26.8",
-    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
+    "@babel/plugin-transform-object-rest-spread": "^7.25.9",
     "@babel/plugin-transform-react-jsx": "^7.25.9",
-    "babel-loader": "^8.4.1",
-    "clean-webpack-plugin": "^2.0.2",
-    "copy-webpack-plugin": "^5.1.2",
-    "css-loader": "^6.11.0",
-    "file-loader": "^5.1.0",
-    "nodemon": "^2.0.22",
-    "style-loader": "^1.3.0",
+    "babel-loader": "^9.2.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^12.0.2",
+    "css-loader": "^7.1.2",
+    "file-loader": "^6.2.0",
+    "nodemon": "^3.1.9",
+    "style-loader": "^4.0.0",
     "webpack": "^5.97.1",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^6.0.1"
   },
   "dependencies": {
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
-  },
-  "resolutions": {
-    "acorn": "npm:acorn-with-stage3"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   }
 }
