@@ -6,20 +6,20 @@ import { Demos } from "./panels/Demos.jsx";
 import { entrypoints } from "uxp";
 
 const demosController = new PanelController(() => <Demos />, {
-    id: "demos"
+  id: "demos"
 });
 
 
 entrypoints.setup({
-    plugin: {
-        create(plugin) {
+  plugin: {
+    create (plugin) {
             /* optional */ console.log("created", plugin);
-        },
-        destroy() {
-            /* optional */ console.log("destroyed");
-        }
     },
-    panels: {
-        demos: demosController
+    destroy () {
+            /* optional */ console.log("destroyed");
     }
+  },
+  panels: {
+    demos: demosController
+  }
 });
